requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-froala-editor==4.0.10
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
